HYD-5441 Use Artifactory in place of chroma-externals for JS modules
Background


When building / running tests, we use chroma-externals to host our
dependencies. The dependencies go through a process like this:

1. Locally store the flattened dependency tree into chroma-externals using
ziplocker and commit to git (after review).
2. On Jenkins, fetch the externals and rebuild them using ziplocker for the
host arch. This is done because some modules use C or C++ and need to be
compiled.
3. Use the resulting dep tree to ship or test.

This has worked, but we can now cut out all these steps. In the 3.0
release, we use Artifactory as an npm registry. Instead of checking deps
into chroma-externals we can pull from Artifactory from Jenkins. This
has the following benefits:

1. No need to run ziplocker before checkin / after checkout. Repo should be
able to be deleted.
2. Easier gerrit pushes / reviews. ui-modules bumps should only effect
package.json
3. Faster builds / tests.

Implementation

1. Remove ziplocker specific code from chroma repo.
2. Update Jenkins scripts to fetch npm modules from artifactory.
3. Use npm shrinkwrap over ui-modules to lock down dependency tree.

  - Remove ui-modules deps from chroma-externals.
  - Switch to using Artifactory for fetches.

Change-Id: Ic4d617ccebf7b0a232d6cc4a5fa5473ec7fb62ca
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/18840
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>