HYD-5592 Starting copytools fails, no /sbin/start
The way in which the daemon's for the copytool are are
started is not EL7 compatible. Using /sbin/start we
need to upgrade these for 3.0 to use more regular service
mechanisms.  This is the minimal set of reasonable changes
that I could find.

Create EL6 Style service files for the copytool and
copytool-monitor control. Update the agent to use this
using the service control. This required extending the
service control to allow additional arguments to be passed.

 - Extend service control, start, stop, status and restart
   to include the ability to accept additional arguments
 - Add tests for the above.
 - Remove reload from service control because it would need
   to be extended for arguments but was already broken and
   not used yet.
 - Update users of service control.
 - Add init files for copytool and copytool-monitor
 - Update Makefiles
 - Update manage_copytool to use new services
 - Remove now unused command copytool_vars from chroma_agent

Change-Id: I6652c1edf1bd6a6dafe8fdcf9a957b3b5ba39cc7
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18857
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>