HYD-5607 AlertEvent records do not have message field populated
New AlertEvents do not have a message field or an alert item.

A change to make the AlertEvents have a beginning and an end was
incorrect and didn't save the record. In the medium term we will
place around with this code to make the AlertEvents be specific
types for the AlertEvent in question. For this reason the solution
here is to simple pass all the elements of the AlertEvent into
the register_event routine. A better solution but more change at
a bad time would be to return not an AlertEvent but a descriptor
of how to create an AlertEvent.

 - Add message_str and alert to the register_event routine.
 - Add some very basic additional tests to check that AlertEvents
   are created and that the basic attributes are correct.

Change-Id: I525c3e44e6e017f856f62cf0055cae342a1ad117
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18864
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>