HYD-5634 Gather help for the LU team when Lustre unmount hangs
When we get Lustre unmount hangs, such as LDEV-323, one useful piece
of information the Lustre team could use is a stack trace of all of
the kernel threads.

So let's try to gather this for them.

Also, let's try to recover the node so that it can be used for
remaining tests.

 - get kernel thread stack trace when umount hangs
 - reset the node so that remaining tests have a chance of passing

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Ib51be933be5affc74787711c319fc8fb62370243
Reviewed-on: http://review.whamcloud.com/19288
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>