HYD-5636 Update upgrade testing to upgrade from 2.4
Since 2.3.x.x does not run on EL6 > 6.6 we need to update the upgrade
testing to start with 2.4.x.x which includes it's own pacemaker and so
will still install on EL6.

This is needed to be able to test with RHEL 6 since it is effectively currently 6.7 (6.8 soon) for every installation.

 - update the chroma-blessed build number to use 2.4.1.2
 - provide an information post-upgrade audit of the RPM inventory on all
   of the nodes

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I40d0a9b0788fabea684e77bad83eb4b20589aaab
Reviewed-on: http://review.whamcloud.com/18877
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>