HYD-5661 Long polling has a race condition
Long polling was using the save and delete callbacks from
Django to indicate that the database had changed. However
in cases where transactions are used the changes will not
be visible to other processes (or cursors). This means a
pending query would read 'old' data and not receive a
notification of the 'new' data.

The solution is to pend the long polling notifications until
the commit happens. Not so easy because Django provide no
callback for the commit. The solution therefore is to
chimpanzee patch the commit so that commits do provide a
notification. The patch is restored once a commit or rollback
happens.  One performance advantage of this is that many
changes often roll into a single transaction and so these
can be changed into a single rpc notification to the
job_scheduler. This is probably a small saving as most
changes occur in the job_scheduler.

Tests have been added to cover the new code and generally
improve the long polling testing. Logging has also been
added.

 - Add batching of changes that occur within a transaction
 - Add unit tests to cover new functionality
 - Rename table_change to tables_changed as the function
   now takes a list of tables.

Change-Id: Ifc8a5998f1b019cb7023fc4af85fb8031bf80ffd
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19018
Tested-by: Jenkins
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>