HYD-5684 Race condition causes transition to monitored mode to fail.
During the installation in monitored mode that LNet configuration is
transitioned to 'LNet Unloaded' so that monitoring can begin. If the
state transitions to 'LNet Stopped' or 'LNet Started' before the command
completes then the final transition to 'Monitored' fails is cancelled because
the state of LNet is not in the required 'State'. This failure mode was
intermittent but made a certainty by the introduction of another slower
dependency (configuration of rsyslog) which meant the transition to another
non 'LNet Unloaded' state always occurred. The solution is to add acceptable
states of the required LNet Configuration so that once the LNet is being
monitored any transition is fine.

As part of this ticket a fix for a problem in LearnEvent is added. The
implementation of the Django class GenericForeignKey is strange because
it expects to write a ContentType to the database but on a read it expects
to receive an index in the form of an int. This had been missed in the
implementation. The solution is to update VariantGenericForeignKey and
LearnedEvent to deal with this anomaly. A follow on patch will simplify this
code further still, simplifying both  VariantGenericForeignKey and LearnedEvent.

 - Add acceptable methods as all operation LNet States for the transition to
   Monitored state.
 - Update LearnedEvent and VariantGenericForeignKey to deal with the
   inconsistent way that the parent class GenericForeignKey uses the DB.

Tests and simplification to be added as a follow on. This ticket review excepting
will be landed as is to allow for HP2 release.

Change-Id: Ieda3f101c0c6a811bec34f43c1ba1b44239894a1
Reviewed-on: http://review.whamcloud.com/19030
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>