HYD-5521 Make test's cibadmin robust
In RHEL 6.8, pacemaker seems to take longer to settle after being
restarted than it did previously.  This was tripping up the simple
use of cibadmin in test as it was not retrying on errors expected
when querying pacemaker shortly after it is restarted.

Make the test implentation/use of cibadmin as robust as the product's.

 - retry on expected temporary errors from pacemaker, such as
   what happens when querying pacemaker right after it is restarted

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I139baf3b46152ce758b0457d06ba5999c5084144
Reviewed-on: http://review.whamcloud.com/19482
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>