HYD-5690 Raising IML plugin alerts gives
Raising an Alert of type StorageResourceRecord causes
the /api/storage_resource endpoint to fail with an error
because to_dict was removed from alert_base. This quick
patch replaces the removed code, but does not cover the
test hole. A follow on patch will do that.

The fix is to add do_dict back into the Alert base class.
It is not quite like for like because end used to == begin
until the alert is complete. However no end is None until the
alert is complete and so until end is not none we return being.

 - Add to_dict back into AlertStateBase

Change-Id: I565477b4bd8c6f5ec89e01f0372b31d9d6cc9637
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19065
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>