HYD-5447 Filesystem not starting after OST fails to register
Testing was showing filesystem was not starting because one or
more OSTs were failing to register.

The mount command was failing at the first attempt but succeeding
after a second atempt (made due to HYD-1040). The second attempt
was not modifying the result variable and so, even though successful
FileSystemLdiskfs.mount() raises an exception.

  - update result with that of the second mount attempt

Change-Id: I883f59a8b5fc530054e4f94b7c4fafaf5cfd1686
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19091
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>