HYD-5447 Add filesystem ldiskfs mount tests
Testing was showing filesystem was not starting because one or
more OSTs were failing to register.

The mount command was failing at the first attempt but succeeding
after a second atempt (made due to HYD-1040). The second attempt
was not modifying the result variable and so, even though successful
FileSystemLdiskfs.mount() raises an exception.

This is a follow-on patch containing the unit tests

  - add unit tests for ldisk mount command

Change-Id: I7f1c371a5d8ba78c269f99ffa7302f80213a3819
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19114
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>