HYD-5736 Error updating chroma alertstate
When doing an update_or_create on resource with a different,
non-error alert state to a resource that has an alert state
triggered in the GUI, plugin_runner.log errors are found.

The cause of this is that the resource manager will delete
ResourceRecordAlerts and these may be referenced by AlertEmails
if emailing of alerts it enabled. The options are to delete
the record the emails were sent, keep all the records in the
resource storage code or create a 'break' in the chain.

This change stops deleting the ResourceRecordAlerts and then
as a general solution stores the message of the alert in a
persistant manner so that the references to the ResourceRecords
do not need to exist. Allowing for there deletion. A positive
knock on of this is that messages for all alerts are created
and stored meaning that alert fetches will be more efficient
and the alerts can be searched for matches in the message (this
may require gui changes).

 - Refactor AlertBase and children to store the message on the
   first fetch
 - For StorageResourceRecords carry out a fetch immediatly after
   creation so that the persistance is 'instant'
 - Add tests to ensure StorageResources are better tested.

Change-Id: Icf662012be715b9aeb559cbe5d9e72a5dca0db13
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19201
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>