HYD-5630 STONITH resource not getting created
STONITH and other configuration is not always carried
out on nodes.

There is a code path (could be more than one) that would allow this.

Two Nodes A and B

Node A runs configure_pacemaker but B is the dc so it does not
configure pacemaker.

The dc switches from B to A (this is a perfectly possible scenario)

Node B runs configure_pacemaker but A is the dc so it does not
configure pacemaker.

Hence Pacemaker is not configured. This scenario is quite likely be
imeadiatly before configuring pacemaker we restart pacemaker (this
is not new) which would tend to cause the DC to switch to the other
node.

The solution is for the configure_pacemaker routines to not exit until
pacemaker is configured. So the logic is

while pacemaker_not_configured:
  if I_am_the_dc:
    configure_pacemaker
  else:
    wait a little

With the addition of timeouts this would seem to provide the state
where the pacemaker must be configured for the method to report success

 - Update manage_pacemaker to use the new routines and make better use
   of the pc class
 - Update pacemaker class to contain more of the pacemaker logic (cibadmin
   in particular
 - Add/modify tests to test changes

Change-Id: I070db8e1265eb724054f79ee665b2217528c21fa
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19406
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>