HYD-5765 ui-modules: Bump intel-gui to 4.4.4
Now the HYD-5720 has landed we need to update intel-gui in the 
ui-modules directory to use version 4.4.4.

This patch will update ui-modules/package.json such that the version 
of intel-gui is 4.4.4. HYD-5720 is the one and only change going into 
this patch, as is customary when we list them in this ticket.

Here is the ticket info for HYD-5720:

The /storage_resource endpoint contains records that do not 
have the available_actions field that other resources do.

Because of this, the actions dropdown on an available 
alert will throw an error when someone tries to access it.

We should either not render the actions dropdown if there 
are no available alerts, or figure out if it makes sense for 
storage_resource records to contain actions.

  - update intel-gui to 4.4.4 which includes HYD-5720

Change-Id: I0cbd2222bc224eda2e4d386b2f36f7f17235f6ca 
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Change-Id: I0cbd2222bc224eda2e4d386b2f36f7f17235f6ca
Reviewed-on: http://review.whamcloud.com/19246
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>