HYD-5778 Check to see if client is mounted fails on EL7
During the remote operations test to detect a successful mount
operation, the output of the 'mount' command is searched to find an
expected string.  This is failing in el7 efs tests.

This output has changed between el6 and el7 as in el7 it is the
/etc/mtab file but in el7, /etc/mtab is a symlink to
/proc/self/mounts. This results in subtle differences in the output
of the 'mount' command on the different distros.

  - just search for the ' on /mnt/<fs_name> ' (mount point) string in
    mount output and not the mount device path

Change-Id: I2a581fe34c6a380d8277bf76d96d40ce9ab72788
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19333
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>