HYD-5779 Fix incorrect call to CommandExecutionError
The CommandExecutionError was changed to require a RunResult and the
command arguments rather than rc,stdout,stderr and the command 
arguments. In a non-happy code path in cibadmin however the old
arguments were still being used.

Changing cibadmin had the knock on effect of requiring the calls that
generate the data to naturally return a RunResult rather than three 
arguments of rc,stdout,stderr in a tuple. The changes to do this are
limited to the cibadmin command and the 4 places that consume the result.


  - Supply AgentShell.RunResult namedtuple as argument to
    CommandExecutionError in cibadmin method, return RunResult
  - Change usages of cibadmin to handle RunResult on return

Change-Id: I1fa56493439d6bf2d4f536339c8dbf0a13f29774
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19335
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>