HYD-5751 Storage plugin UpperBoundCondition always triggered
In the test plugin the upperbound is always triggered because the
bound test compares the value as a string to the bound as an int.
In Python ('0' > 100) == True. Because the data is pushed from
an outside source via json the best solution is to appropriately
cast it upon arrival to the correct type. This change also allows
a simplication of the use of passwords by generalising the
encrypt functionality so that all types encrypt but the default
encryption is a null task.

A final change is to allow the default value to be used in the
case of missing values from the rest api.

 - Update storage_resource to apply defaults for missing values
 - Update attributes to base_resource_attribute to provide a
   standard entrpt function and a cast function for all types
 - Add cast tests for each attribute value

Change-Id: Id6b84b3ee526a6768e57e3447f2cf6efdf2d62e5
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19376
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>