HYD-5366 On EL7 HTTP Agent throws SystemError on startup
Running on EL7 HTTP Agent throws the following error

 SystemError NULL result without error in PyObject_Call

An attempt to fix this by removing optimization fixed the issue
with el6.7 but the issue has returned with el7. Upgrading to
0.4.0-1 is expected to fix this, and so an upgrade to 0.4.2 is
a pertinent change at this point.

Upgrade python-gevent to 0.4.2 and revert the optimization removal
so the library is now compiled as distributed.

 - Remove changes to spec file
 - Upgrade to 0.4.2 source

Change-Id: I960637137b08f1d19e0c46af6757c28a0c62d27e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19350
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>