HYD-5817 On el7 get_resource_running fails because pacemaker is using FQDN names
On el7 get_resource_running because pacemaker is using FQDN names on el7
but on el6 it uses nodenames. This means the node name given to crm_attribute
must be varied.

The solution is to detect the version of the remote node and the -N parameter
given to crm_attribute. The addition of a distro_version method in RemoteOperations
allows for easy detection.

 - Change the get_resource_running to use nodename or fqdn as appropriate.
 - Add distro_version to RemoteOperations so that the detection of distro version
   is easy. This routine works for config hosts and for api hosts.

Change-Id: Iee25188d4ef2784786e6efd600f63465e6baa167
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19404
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>