HYD-5790 On el7 check_ha_config fails because pacemaker is returning FQDN names
On el7 check_ha_config fails because pacemaker is returning FQDN names on el7
but on el6 it used to return nodenames.

The solution is to expand all returned names to their FQDN before doing any
comparisons. Whilst exploring this error I laid some of the code in
check_ha_config slightly differently for clarity and also changed the names
of two variables in local functions which were shadowing variables in the
parent function.

 - Change the comparison of the return data from cibadmin to always compare
   FQDN with FQDN
 - For clarity and safety layout (but not not functionally change) the code
   in check_ha_config and it local functions more clearly and change two
   variable names to removing shadowing

Change-Id: Ib94a5c7808588baa626d9094b60eab1ecadf864e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19352
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>