HYD-5802 Removing a server with OST(s) fails
Removing a server with targets on the server fails because Pacemaker is
stopped before the non-active ManagedTargets are unconfigured.

The issue is that inactive-ManagedTargets are not dependent on Pacemaker
being started and so are not unconfigured before it is stopped. On the
other hand active-ManagedTargets are dependent and so the unconfiguration
happens in the correct order before Pacemaker is stopped.

However this causes problems because when Removing a host Pacemaker is
stopped before the ManagedTarget is unconfigured, so to counter this it
is a dependency that for Pacemaker to be unconfigured as host must be
removed from the filesystem.

The solution is to add dependency for Pacemaker along side the existing
dependency for LNet in the code.

Finally because it is no possible to remove a node form the filesystem
by unconfiguring Pacemaker the message has been updated to be clearer
that the Target is being removed from the Filesystem - rather than Target
is being removed from the configuration. Finally it has been made the
case that Unconfigure Pacemaker/Corosync is not offered where a host has
a ManagedTargetMount associated with it.

This patch doesn't close the test hole that obviously exists here for the
simple reason that no tests appear to exist to augment with this small change.
So for pragmatic time reasons, we are past the planned rc date, this
lack of automated testing will not be resolved. This is not a regression and
so is acceptable.

 - Add Pacemaker dependency to mirror the LNet dependency for the inactive
   case of a ManagedHost
 - Add Pacemaker dependency that requires a ManagedHost to be removed before
   the Pacemaker configuration is unconfigured.
 - Change can_run for UnconfigureCorosyncJob and UnconfigurePacemakerJob
   so that unconfigure Pacemaker/Corosync is not presented when the host
   has a ManagedTargetMount associated with it.

Change-Id: Ic69606806bc014056ab158adc199ece1008999b6
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19492
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>