HYD-5897 Some unit tests fail running a debugger which is annoying.
Because PowerControlTestCase presumes the threads that will be
running to detect unexpected threads, rather than recording the
threads at the start of a test and at the end of a test, the test
will fail if run in a debugger that will quite reasonably have
additional tests running.

Change the detection mechanism to use the delta between threads at the
beginning and threads at the end.

 - Add detection to test_power_control tests.

Change-Id: Idc62fd3143ed815d2291a5993736419422632834
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19373
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>