HYD-5857 Typo in HYD-5817 where FQDN was labeled as FDQN
The implementation of HYD-5817 used FDQN as a dict accessor
rather than FQDN. This error was not spotted in test because
it is used only on the EL7 path which was not a test pass 
required for landing.
 
Correct FDQN to FQDN and get_resource_running will work on EL6 
and EL7. This fact has been tested on both distributions.
 
 - Update get_resource_running to used FQDN in the dictionary.

Change-Id: I885db5f90347ceb4b166e29aeae04e1324906dca
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19478
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>