HYD-5827 Alert messages do not change when the AlertType changes
The solution to HYD-5736 caused Alert messages to be cached rather
than regenerated. This made it possible for the GUI to search them
but also made it possible for the AlertItems referenced to be deleted
with the message becoming impossible to generate.

However when an Alert is cast to a different Alert for example
CommandRunningAlert becomes CommandCancelledAlert the message must
be regenerated or it may well be incorrect.

The cast function for Alert's needs to regenerate the message after
the alert is cast.

 - Regenerate the Alert message in the AlertStateBase.cast method
 - Add a new test_alert.py in the unit tests to verify this change.

Change-Id: Ide3bc8c5cfab67dfacd6b77dc0414b594a164612
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19483
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>