HYD-1943 Add help info into chroma-config command
It would be more intuitive or customary to support “-h” and "-help".
As would a chroma-config MAN page, to give more information
about the command.

To do this add the “-h” and "-help" options to service_config.py
and write a man page and add it to the chroma-manger.spec file

 - Edit service_config.py to have “-h” and "-help" options.
 - Edit chroma-manager.spec to add man page file onto user system
 - Edit chroma-manger.spec file to run makewhatis

Change-Id: Idc1bd3b23455c24fdf2a6573bb32c6e912fb82a3
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/19649
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>