HYD-5492 Virtualenv --no-site-packages option is now deprecated
The virtualenv python package for creating virtual python (isolated)
environments is used extensively in IML testing. It has an option
'--no-site-packages' which is used in IML but which is now deprecated
and could be removed in any future release of virtualenv package.

Usage of the option should be removed from all versioned code in
order to avoid the occurrence of related bugs in the future.

--no-site-packages has been the default behaviour since version 1.7

 - Remove --no-site-packages from cluster_setup

Change-Id: I69f76404cb4974a7256fdb2f837070897577a0ae
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/19506
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>