HYD-5417 Logrotate Compress scope if global for chroma-agent*.log
Because the compress keyword is outside of the main stanza, it applies
to all log files that logrotate handles (logrotate just does an include of
everything in /etc/logrotate.d, it doesn't care what the files in
/etc/logrotate.d are named).

This causes all log files of affected systems to be compressed, regardless
of the site's intention to compress or not. We should only 
compress chroma-agent.log files, not all system logs.

 - Edit config file so compress setting is within the scope for 
   /var/log/chroma-agent*.log

Change-Id: Ibe882a099ccd616a08df409d58a03e7eb499f5f4
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/19508
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>