HYD-5840 Systemctl needs to be told to load new systemd init files
When creating a systemd init file as we do today for el7 and copytool,
the systemctl agent needs to be told to load them. This is done by
executing systemctl daemon-reload. Because this was missing from the
copytool implementation it failed to work as planned.

Add the reload functionality to the ServiceControl class and add
appropriate tests. In start_monitored_copytool and
stop_monitored_copytool issue the daemon_reload command after
creation or deletion.

 - Add daemon_reload as a class method for ServiceControl
 - Add daemon_reload as a non-action for ServiceControlEL6
 - Add daemon_reload as a action for ServiceControlEL7
 - Add tests for ServiceControlEL6 and ServiceControlEL7
 - Add calls to daemon_reload after creation and deletion of
   init files

Change-Id: I8a786b7f7e7013e3aa688e43a0efffb2c1bcf950
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19514
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>