HYD-5844 Add .gitlint file to chroma
Add a .gitlint file to the root of chroma. This file will have some
default rules that a user can run with the gitlint
(http://jorisroovers.github.io/gitlint/) tool.

For now, this tool will need to be run manually because it uses
OrderedDict which is not supported in Python 2.6

Once we exclusively use Python 2.7 we can switch over, but for now this
at least gives some automated feedback a developer can run before push,
or add to a hook if they only use Python 2.7

  - Add .gitlint file.

Change-Id: I9709fb96b2867633a2ba4bf25b47c1416bfeec4b
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/19519
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>