HYD-5856 Corosync2 configuration incorrectly counts the corosync peers
Corosync 2 configuration has a method called corosync peers which uses
the mcastport information to calculate the peers of the requested node.
Because database updates are asynchronous it needs to track the current
configurations in realtime and respond to additions and deletions of
hosts.

The deletions code has a fault in that it does not consider a node that
exists but does not yet have an mcast_port to exist which means that it
will delete it from the current list.
The solution is to remove the filter for (mast_port is not None) and
explicitly check for that in Python meaning we create a list of all
current nodes and all fqdns with an mcast_port.

 - Fetch all the current CorosyncConfigurations rather than a filtered
   list and then filter in Python whilst also building a complete list of
   existing configurations.

Change-Id: If75e585742dde740b0a7929c5807974928062dfd
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19559
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>