HYD-5631 Parse crm_mon for still running stopped targets
Currently in get_resource_locations() we don't account for status lines of
the format:
MGS_424f74  (ocf::chroma:Target):   (target-role:Stopped) Started lotus-13vm6
and when we do encounter one of those, we end up returning that such a
resource is running on host Started.

Update the boilerplate code to change the copyright data to 2016.

Enhance the parsing to account for this.

 - also look for Stopped but still Started resources
 - Update boilerplate code

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I5be9457022824c0c6dbf5300e19f3077473bec93
Reviewed-on: http://review.whamcloud.com/19553
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>