HYD-5855 Skip throws exception because config.get given default parameter
SSI tests running on el7 fail in skipIf decorator in test_updates.py

Python 2.7 will throw an exception if config.get is given a second,
default parameter == False. This error is not experienced with python
2.6 on el6 distributions.

This behaviour is seen because the default version of configparser
provided with python 2.7 takes "option" as the second argument, and
False is not a valid "option".

See https://docs.python.org/2/library/configparser.html for more
details.

  - remove the default 'False' parameter from config.get, the method
    is evaluated as False if the given attribute is missing from the
    config, this is its implicit behaviour
  - change the import of skipIf to be from the module
    django.utils.unittest rather than django.utils.unittest.case
    to be consistent with other integration tests

Change-Id: I87658c70fb987282f83ff19878466aacd3231c46
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19567
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>