HYD-4775 Remove need for chroma-externals pragma
The chroma-externals pragma works well but has an overhead that
when a number of patches exist after a chroma-externals pragma
they must all have the commit message change.

We can change to use the sha1 in the original commit to find
the gerrit commit to pull.

Process is

See if git submodule update works
If not get the sha1 that we require from the error message, this
is not pretty but I couldn't work out how to do it another way.
query git to findout whic commit this sha1 is.
Fetch that commit from gerrit.
Finally to ensure all is well git submodule update should now work.

 - New update_chroma-externals script.
 - Change calls to use the new script.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I651caca7b56af02e39e711fa4d09d88502971e42
Reviewed-on: http://review.whamcloud.com/19560
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>